refactor(match2): Split MatchGroup/Input into two
