+++ components/match2/commons/starcraft_starcraft2/match_group_input_starcraft.lua
@@ -18,7 +18,7 @@ local Table = require('Module:Table')
 local Variables = require('Module:Variables')
 
 local DeprecatedCustomMatchGroupInput = Lua.import('Module:MatchGroup/Input/Starcraft/deprecated')
-local MatchGroupInput = Lua.import('Module:MatchGroup/Input')
+local MatchGroupInputUtil = Lua.import('Module:MatchGroup/Input/Util')
 local OpponentLibraries = require('Module:OpponentLibraries')
 local Opponent = OpponentLibraries.Opponent
 local Streams = Lua.import('Module:Links/Stream')
@@ -47,7 +47,7 @@ function StarcraftMatchGroupInput.processMatch(match, options)
 	Table.mergeInto(match, MatchFunctions.readDate(match.date))
 
 	local opponents = Array.mapIndexes(function(opponentIndex)
-		return MatchGroupInput.readOpponent(match, opponentIndex, OPPONENT_CONFIG)
+		return MatchGroupInputUtil.readOpponent(match, opponentIndex, OPPONENT_CONFIG)
 	end)
 
 	-- TODO: check how we can get rid of this legacy stuff ...
@@ -70,12 +70,12 @@ function StarcraftMatchGroupInput.processMatch(match, options)
 
 	local games = MatchFunctions.extractMaps(match, opponents)
 
-	local autoScoreFunction = MatchGroupInput.canUseAutoScore(match, games)
+	local autoScoreFunction = MatchGroupInputUtil.canUseAutoScore(match, games)
 		and MatchFunctions.calculateMatchScore(games, opponents)
 		or nil
 
 	Array.forEach(opponents, function(opponent, opponentIndex)
-		opponent.score, opponent.status = MatchGroupInput.computeOpponentScore({
+		opponent.score, opponent.status = MatchGroupInputUtil.computeOpponentScore({
 			walkover = match.walkover,
 			winner = match.winner,
 			opponentIndex = opponentIndex,
@@ -93,19 +93,19 @@ function StarcraftMatchGroupInput.processMatch(match, options)
 
 	local winnerInput = match.winner --[[@as string?]]
 	local finishedInput = match.finished --[[@as string?]]
-	match.finished = MatchGroupInput.matchIsFinished(match, opponents)
+	match.finished = MatchGroupInputUtil.matchIsFinished(match, opponents)
 
 	if match.finished then
-		match.resulttype = MatchGroupInput.getResultType(winnerInput, finishedInput, opponents)
-		match.walkover = MatchGroupInput.getWalkover(match.resulttype, opponents)
-		match.winner = MatchGroupInput.getWinner(match.resulttype, winnerInput, opponents)
-		MatchGroupInput.setPlacement(opponents, match.winner, 1, 2)
-	elseif MatchGroupInput.isNotPlayed(winnerInput, finishedInput) then
-		match.resulttype = MatchGroupInput.getResultType(winnerInput, finishedInput, opponents)
+		match.resulttype = MatchGroupInputUtil.getResultType(winnerInput, finishedInput, opponents)
+		match.walkover = MatchGroupInputUtil.getWalkover(match.resulttype, opponents)
+		match.winner = MatchGroupInputUtil.getWinner(match.resulttype, winnerInput, opponents)
+		MatchGroupInputUtil.setPlacement(opponents, match.winner, 1, 2)
+	elseif MatchGroupInputUtil.isNotPlayed(winnerInput, finishedInput) then
+		match.resulttype = MatchGroupInputUtil.getResultType(winnerInput, finishedInput, opponents)
 		match.winner = nil
 	end
 
-	MatchGroupInput.getCommonTournamentVars(match)
+	MatchGroupInputUtil.getCommonTournamentVars(match)
 
 	match.stream = Streams.processStreams(match)
 	match.vod = Logic.nilIfEmpty(match.vod)
@@ -121,7 +121,7 @@ end
 ---@param dateInput string?
 ---@return {date: string, dateexact: boolean, timestamp: integer, timezoneId: string?, timezoneOffset: string?}
 function MatchFunctions.readDate(dateInput)
-	local dateProps = MatchGroupInput.readDate(dateInput, {
+	local dateProps = MatchGroupInputUtil.readDate(dateInput, {
 		'matchDate',
 		'tournament_startdate',
 		'tournament_enddate',
@@ -162,7 +162,7 @@ end
 ---@return fun(opponentIndex: integer): integer?
 function MatchFunctions.calculateMatchScore(maps, opponents)
 	return function(opponentIndex)
-		local calculatedScore = MatchGroupInput.computeMatchScoreFromMapWinners(maps, opponentIndex)
+		local calculatedScore = MatchGroupInputUtil.computeMatchScoreFromMapWinners(maps, opponentIndex)
 		if not calculatedScore then return end
 		local opponent = opponents[opponentIndex]
 		return calculatedScore + (opponent.extradata.advantage or 0) - (opponent.extradata.penalty or 0)
@@ -224,7 +224,7 @@ end
 ---@return table
 function MatchFunctions.getExtraData(match, numberOfGames)
 	local extradata = {
-		casters = MatchGroupInput.readCasters(match, {noSort = true}),
+		casters = MatchGroupInputUtil.readCasters(match, {noSort = true}),
 		headtohead = tostring(Logic.readBool(Logic.emptyOr(match.headtohead, Variables.varDefault('headtohead')))),
 		ffa = 'false',
 	}
@@ -281,7 +281,7 @@ function MapFunctions.readMap(mapInput, subGroup, opponentCount)
 
 	map.finished = MapFunctions.isFinished(mapInput, opponentCount)
 	local opponentInfo = Array.map(Array.range(1, opponentCount), function(opponentIndex)
-		local score, status = MatchGroupInput.computeOpponentScore({
+		local score, status = MatchGroupInputUtil.computeOpponentScore({
 			walkover = mapInput.walkover,
 			winner = mapInput.winner,
 			opponentIndex = opponentIndex,
@@ -292,10 +292,10 @@ function MapFunctions.readMap(mapInput, subGroup, opponentCount)
 
 	map.scores = Array.map(opponentInfo, Operator.property('score'))
 
-	if map.finished or MatchGroupInput.isNotPlayed(map.winner, mapInput.finished) then
-		map.resulttype = MatchGroupInput.getResultType(mapInput.winner, mapInput.finished, opponentInfo)
-		map.walkover = MatchGroupInput.getWalkover(map.resulttype, opponentInfo)
-		map.winner = MatchGroupInput.getWinner(map.resulttype, mapInput.winner, opponentInfo)
+	if map.finished or MatchGroupInputUtil.isNotPlayed(map.winner, mapInput.finished) then
+		map.resulttype = MatchGroupInputUtil.getResultType(mapInput.winner, mapInput.finished, opponentInfo)
+		map.walkover = MatchGroupInputUtil.getWalkover(map.resulttype, opponentInfo)
+		map.winner = MatchGroupInputUtil.getWinner(map.resulttype, mapInput.winner, opponentInfo)
 	end
 
 	return map, subGroup
@@ -305,7 +305,7 @@ end
 ---@param opponentCount integer
 ---@return boolean
 function MapFunctions.isFinished(mapInput, opponentCount)
-	if MatchGroupInput.isNotPlayed(mapInput.winner, mapInput.finished) then
+	if MatchGroupInputUtil.isNotPlayed(mapInput.winner, mapInput.finished) then
 		return true
 	end
