+++ components/match2/wikis/brawlstars/match_group_input_custom.lua
@@ -17,7 +17,7 @@ local Variables = require('Module:Variables')
 local Streams = require('Module:Links/Stream')
 local BrawlerNames = mw.loadData('Module:BrawlerNames')
 
-local MatchGroupInput = Lua.import('Module:MatchGroup/Input')
+local MatchGroupInputUtil = Lua.import('Module:MatchGroup/Input/Util')
 
 local FIRST_PICK_CONVERSION = {
 	blue = 1,
@@ -42,19 +42,19 @@ function CustomMatchGroupInput.processMatch(match, options)
 	local finishedInput = match.finished --[[@as string?]]
 	local winnerInput = match.winner --[[@as string?]]
 
-	Table.mergeInto(match, MatchGroupInput.readDate(match.date, {'tournament_enddate'}))
+	Table.mergeInto(match, MatchGroupInputUtil.readDate(match.date, {'tournament_enddate'}))
 
 	local opponents = Array.mapIndexes(function(opponentIndex)
-		return MatchGroupInput.readOpponent(match, opponentIndex, {})
+		return MatchGroupInputUtil.readOpponent(match, opponentIndex, {})
 	end)
 	local games = CustomMatchGroupInput.extractMaps(match, #opponents)
 	match.bestof = MatchFunctions.getBestOf(match)
 
-	local autoScoreFunction = MatchGroupInput.canUseAutoScore(match, games)
+	local autoScoreFunction = MatchGroupInputUtil.canUseAutoScore(match, games)
 		and MatchFunctions.calculateMatchScore(games, match.bestof)
 		or nil
 	Array.forEach(opponents, function(opponent, opponentIndex)
-		opponent.score, opponent.status = MatchGroupInput.computeOpponentScore({
+		opponent.score, opponent.status = MatchGroupInputUtil.computeOpponentScore({
 			walkover = match.walkover,
 			winner = match.winner,
 			opponentIndex = opponentIndex,
@@ -62,15 +62,15 @@ function CustomMatchGroupInput.processMatch(match, options)
 		}, autoScoreFunction)
 	end)
 
-	match.finished = MatchGroupInput.matchIsFinished(match, opponents)
+	match.finished = MatchGroupInputUtil.matchIsFinished(match, opponents)
 
 	if match.finished then
-		match.resulttype = MatchGroupInput.getResultType(winnerInput, finishedInput, opponents)
-		match.walkover = MatchGroupInput.getWalkover(match.resulttype, opponents)
-		match.winner = MatchGroupInput.getWinner(match.resulttype, winnerInput, opponents)
-		MatchGroupInput.setPlacement(opponents, match.winner, 1, 2)
-	elseif MatchGroupInput.isNotPlayed(winnerInput, finishedInput) then
-		match.resulttype = MatchGroupInput.getResultType(winnerInput, finishedInput, opponents)
+		match.resulttype = MatchGroupInputUtil.getResultType(winnerInput, finishedInput, opponents)
+		match.walkover = MatchGroupInputUtil.getWalkover(match.resulttype, opponents)
+		match.winner = MatchGroupInputUtil.getWinner(match.resulttype, winnerInput, opponents)
+		MatchGroupInputUtil.setPlacement(opponents, match.winner, 1, 2)
+	elseif MatchGroupInputUtil.isNotPlayed(winnerInput, finishedInput) then
+		match.resulttype = MatchGroupInputUtil.getResultType(winnerInput, finishedInput, opponents)
 		match.winner = nil
 	end
 
@@ -101,7 +101,7 @@ function CustomMatchGroupInput.extractMaps(match, opponentCount)
 		map.extradata = MapFunctions.getExtraData(map, opponentCount)
 
 		local opponentInfo = Array.map(Array.range(1, opponentCount), function(opponentIndex)
-			local score, status = MatchGroupInput.computeOpponentScore({
+			local score, status = MatchGroupInputUtil.computeOpponentScore({
 				walkover = map.walkover,
 				winner = map.winner,
 				opponentIndex = opponentIndex,
@@ -110,13 +110,13 @@ function CustomMatchGroupInput.extractMaps(match, opponentCount)
 			return {score = score, status = status}
 		end)
 
-		map.finished = MatchGroupInput.mapIsFinished(map, opponentInfo)
+		map.finished = MatchGroupInputUtil.mapIsFinished(map, opponentInfo)
 
 		map.scores = Array.map(opponentInfo, Operator.property('score'))
-		if map.finished or MatchGroupInput.isNotPlayed(map.winner, finishedInput) then
-			map.resulttype = MatchGroupInput.getResultType(winnerInput, finishedInput, opponentInfo)
-			map.walkover = MatchGroupInput.getWalkover(map.resulttype, opponentInfo)
-			map.winner = MatchGroupInput.getWinner(map.resulttype, winnerInput, opponentInfo)
+		if map.finished or MatchGroupInputUtil.isNotPlayed(map.winner, finishedInput) then
+			map.resulttype = MatchGroupInputUtil.getResultType(winnerInput, finishedInput, opponentInfo)
+			map.walkover = MatchGroupInputUtil.getWalkover(map.resulttype, opponentInfo)
+			map.winner = MatchGroupInputUtil.getWinner(map.resulttype, winnerInput, opponentInfo)
 		end
 
 		table.insert(maps, map)
@@ -137,7 +137,7 @@ CustomMatchGroupInput.processMap = FnUtil.identity
 ---@return fun(opponentIndex: integer): integer?
 function MatchFunctions.calculateMatchScore(maps, bestOf)
 	return function(opponentIndex)
-		return MatchGroupInput.computeMatchScoreFromMapWinners(maps, opponentIndex)
+		return MatchGroupInputUtil.computeMatchScoreFromMapWinners(maps, opponentIndex)
 	end
 end
 
@@ -153,14 +153,14 @@ end
 ---@return table
 function MatchFunctions.getTournamentVars(match)
 	match.mode = Logic.emptyOr(match.mode, Variables.varDefault('tournament_mode', 'team'))
-	return MatchGroupInput.getCommonTournamentVars(match)
+	return MatchGroupInputUtil.getCommonTournamentVars(match)
 end
 
 ---@param match table
 ---@return table
 function MatchFunctions.getExtraData(match)
 	return {
-		mvp = MatchGroupInput.readMvp(match),
+		mvp = MatchGroupInputUtil.readMvp(match),
 	}
 end
 
@@ -189,7 +189,7 @@ function MapFunctions.getExtraData(map, opponentCount)
 	}
 
 	local bans = {}
-	local getCharacterName = FnUtil.curry(MatchGroupInput.getCharacterName, BrawlerNames)
+	local getCharacterName = FnUtil.curry(MatchGroupInputUtil.getCharacterName, BrawlerNames)
 	for opponentIndex = 1, opponentCount do
 		bans['team' .. opponentIndex] = {}
 		for _, ban in Table.iter.pairsByPrefix(map, 't' .. opponentIndex .. 'b') do
@@ -208,7 +208,7 @@ end
 ---@return table
 function MapFunctions.getParticipants(map, opponentCount)
 	local participants = {}
-	local getCharacterName = FnUtil.curry(MatchGroupInput.getCharacterName, BrawlerNames)
+	local getCharacterName = FnUtil.curry(MatchGroupInputUtil.getCharacterName, BrawlerNames)
 	for opponentIndex = 1, opponentCount do
 		for _, player, playerIndex in Table.iter.pairsByPrefix(map, 't' .. opponentIndex .. 'p') do
 			participants[opponentIndex .. '_' .. playerIndex] = {player = player}
