+++ components/match2/wikis/crossfire/match_group_input_custom.lua
@@ -17,7 +17,7 @@ local Variables = require('Module:Variables')
 local DateExt = require('Module:Date/Ext')
 local Streams = require('Module:Links/Stream')
 
-local MatchGroupInput = Lua.import('Module:MatchGroup/Input')
+local MatchGroupInput = Lua.import('Module:MatchGroup/Input/Util')
 
 local NP_STATUSES = {'skip', 'np', 'canceled', 'cancelled'}
 local ALLOWED_STATUSES = { 'W', 'FF', 'DQ', 'L', 'D' }
