+++ components/match2/wikis/osu/match_group_input_custom.lua
@@ -16,7 +16,7 @@ local Variables = require('Module:Variables')
 local DateExt = require('Module:Date/Ext')
 local Streams = require('Module:Links/Stream')
 
-local MatchGroupInput = Lua.import('Module:MatchGroup/Input')
+local MatchGroupInput = Lua.import('Module:MatchGroup/Input/Util')
 
 local NP_STATUSES = {'skip', 'np', 'canceled', 'cancelled'}
 local ALLOWED_VETOES = {'decider', 'pick', 'ban', 'defaultban', 'protect'}
