+++ components/match2/wikis/overwatch/match_group_input_custom.lua
@@ -15,7 +15,7 @@ local Streams = require('Module:Links/Stream')
 local Table = require('Module:Table')
 local Variables = require('Module:Variables')
 
-local MatchGroupInput = Lua.import('Module:MatchGroup/Input')
+local MatchGroupInput = Lua.import('Module:MatchGroup/Input/Util')
 
 local DEFAULT_BESTOF = 3
