+++ components/match2/wikis/rainbowsix/match_group_input_custom.lua
@@ -16,7 +16,7 @@ local Streams = require('Module:Links/Stream')
 local Table = require('Module:Table')
 local Variables = require('Module:Variables')
 
-local MatchGroupInput = Lua.import('Module:MatchGroup/Input')
+local MatchGroupInput = Lua.import('Module:MatchGroup/Input/Util')
 
 local MAX_NUM_BANS = 2
 local DUMMY_MAP = 'null' -- Is set in Template:Map when |map= is empty.
