+++ components/match2/wikis/splatoon/match_group_input_custom.lua
@@ -14,7 +14,7 @@ local Variables = require('Module:Variables')
 local Streams = require('Module:Links/Stream')
 local WeaponNames = mw.loadData('Module:WeaponNames')
 
-local MatchGroupInput = Lua.import('Module:MatchGroup/Input')
+local MatchGroupInput = Lua.import('Module:MatchGroup/Input/Util')
 local Opponent = Lua.import('Module:Opponent')
 
 local STATUS_SCORE = 'S'
