+++ components/match2/wikis/wildrift/match_group_input_custom.lua
@@ -16,7 +16,7 @@ local Table = require('Module:Table')
 local Variables = require('Module:Variables')
 local Streams = require('Module:Links/Stream')
 
-local MatchGroupInput = Lua.import('Module:MatchGroup/Input')
+local MatchGroupInput = Lua.import('Module:MatchGroup/Input/Util')
 
 local MAX_NUM_PLAYERS = 5
 local DEFAULT_BESTOF = 3
