+++ components/match2/wikis/zula/match_group_input_custom.lua
@@ -17,7 +17,7 @@ local DateExt = require('Module:Date/Ext')
 local Streams = require('Module:Links/Stream')
 
 local Opponent = Lua.import('Module:Opponent')
-local MatchGroupInput = Lua.import('Module:MatchGroup/Input')
+local MatchGroupInput = Lua.import('Module:MatchGroup/Input/Util')
 
 local SIDE_DEF = 'ct'
 local SIDE_ATK = 't'
